pa_mqa_logits_fp4: hoist shared helpers into pa_mqa_logits_fp4_common
_i32_buffer / _load_vec4_i32 / _pack_i32_pair_to_i64 / _pack_lo_i64x2_to_i32x8
were duplicated verbatim in the decode and prefill kernels. Move them to one
pa_mqa_logits_fp4_common module and import from both (-88/+63 net, single
source of truth). Drop the now-unused _llvm import from both kernels.

Validated byte-exact (cos=1.0) on decode + prefill op-tests.
